refactor(libutil): convert ActivityType to enum class
Convert the ActivityType enum from old-style `typedef enum` to modern
`enum class` with cleaner value names:

- actUnknown -> ActivityType::Unknown
- actCopyPath -> ActivityType::CopyPath
- actFileTransfer -> ActivityType::FileTransfer
- actRealise -> ActivityType::Realise
- actCopyPaths -> ActivityType::CopyPaths
- actBuilds -> ActivityType::Builds
- actBuild -> ActivityType::Build
- actOptimiseStore -> ActivityType::OptimiseStore
- actVerifyPaths -> ActivityType::VerifyPaths
- actSubstitute -> ActivityType::Substitute
- actQueryPathInfo -> ActivityType::QueryPathInfo
- actPostBuildHook -> ActivityType::PostBuildHook
- actBuildWaiting -> ActivityType::BuildWaiting
- actFetchTree -> ActivityType::FetchTree

Added Logger::Field constructor for ActivityType to support logging.
Added explicit cast for protocol serialization.

+++ src/libstore/daemon.cc
@@ -155,7 +155,8 @@ struct TunnelLogger : public Logger
         }
 
         StringSink buf;
-        buf << STDERR_START_ACTIVITY << act << static_cast<uint64_t>(lvl) << type << s << fields << parent;
+        buf << STDERR_START_ACTIVITY << act << static_cast<uint64_t>(lvl) << static_cast<uint64_t>(type) << s << fields
+            << parent;
         enqueueMsg(buf.s);
     }
